Implement bounded local network discovery

+++ apps/peerdeal_desktop/windows/runner/windows_native_transport.cpp
@@ -639,10 +639,10 @@ WindowsNativeTransport::DecodeFrame(const uint8_t* bytes, std::size_t length) {
       !ReadUint16(bytes, length, &offset, &recipient_length) ||
       !ReadUint32(bytes, length, &offset, &sequence) ||
       !ReadUint32(bytes, length, &offset, &payload_length) || sequence < 1 ||
-      sequence > std::numeric_limits<int32_t>::max() ||
+      sequence > static_cast<uint32_t>(std::numeric_limits<int32_t>::max()) ||
       session_length > kMaxIdBytes || sender_length > kMaxIdBytes ||
       recipient_length > kMaxIdBytes || payload_length < 1 ||
-      payload_length > kMaxPayloadBytes ||
+      payload_length > static_cast<uint32_t>(kMaxPayloadBytes) ||
       length != offset + session_length + sender_length + recipient_length +
                     payload_length) {
     return std::nullopt;
